pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "web3 (>=7.9.0,<8.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "web3-multicall (>=0.0.7,<0.0.8)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false
